Begin the solver rewrite
The current Board format is unwieldy and relies on repeated uses of
solving function and does nothing to address the eventuality of having
to branch the board state to evaluate solution paths.

This rewrite focuses on encapsulating the board state as an object that
can be duplicated as necessary to checkpoint, branch and revert board
states while evaluating solutions.

Testing is also introduced.